admin - added proposal update, assign supervisor, proposal file. teacher management, add new teacher. teacher list

+++ project_management/views.py
@@ -22,13 +22,8 @@ class HomeView(View):
         return render(request, self.template_name,
                       {'courses': running_courses, 'archived': archived_courses})
 
-    @property
-    def is_past_due(self):
-        return datetime.date.today() > self.date
-
-
-def post(self, request, *args, **kwargs):
-    return render(request, self.template_name, {})
+    def post(self, request, *args, **kwargs):
+        return render(request, self.template_name, {})
 
 
 class ProjectDetailsView(LoginRequiredMixin, View):
@@ -70,17 +65,28 @@ class ProjectDetailsView(LoginRequiredMixin, View):
 
     def post(self, request, *args, **kwargs):
         print(request.POST)
-        # delete proposal if the admin tries to delete it
+        # check if it is an ajax call
         if request.is_ajax() and request.user.is_staff:
             print("Came from ajax")
+            delete_proposal = request.POST.get("delete_button")
+            print(delete_proposal)
+
             proposal_id = request.POST.get("proposal_id")
             proposal = Proposal.objects.get(id=proposal_id)
-            proposal.delete()
+            # if the admin is requesting to delete the proposal
+            if delete_proposal == 'true':
+                proposal.delete()
+            # otherwise he is trying to assign a supervisor
+            else:
+                print("Assign Supervisor")
+                proposal.assigned_supervisor = Teacher.objects.get(id=request.POST.get("teacher_id"))
+                proposal.assigned_by = request.user.teacher
+                proposal.save()
+                print(proposal.assigned_supervisor)
 
             course_id = kwargs.get('id')
             course = Course.objects.get(id=course_id)
             filter_by = kwargs.get('filter_by')
-
             if filter_by == 'all':
                 proposals = course.proposal_set.all()
             elif filter_by == 'assigned':
@@ -171,47 +177,25 @@ class UpdateCourseView(LoginRequiredMixin, View):
 class ProposalSubmissionView(LoginRequiredMixin, View):
     template_name = 'project_management/proposal_submission.html'
 
-    context = {
-        'students': Student.objects.all(),
-        'teachers': Teacher.objects.all(),
-        'proposal': '',
-    }
-
     def get(self, request, *args, **kwargs):
         pk = kwargs.get('id')
         course = Course.objects.get(id=pk)
         proposal = Proposal.objects.filter(students=request.user.student, course_id=course.id).first()
 
-        self.context = {
+        context = {
             'students': Student.objects.exclude(proposal__course=course),
             'teachers': Teacher.objects.all(),
             'proposal': proposal
         }
 
-        return render(request, self.template_name, context=self.context)
+        return render(request, self.template_name, context=context)
 
     def post(self, request, *args, **kwargs):
         print(request.POST)
         pk = kwargs.get('id')
         course = Course.objects.get(id=pk)
-        # check if the user requested has permission to delete it and return to proposal page
-        if request.POST.get("delete_button") and request.user.is_staff:
-            print("Came from delete button")
-            for_course = request.POST.get("course_id")
-            student_id = request.POST.get("team_lead_id")
-            student = Student.objects.get(student_id=student_id)
-            proposal = Proposal.objects.get(team_lead=student, course_id=for_course)
-            proposal.delete()
-            # check the redirect url proposal gets deleted but redirection doesn't work
-            return render(request, self.template_name, context=self.context)
 
-        # if it is a normal POST method then create/update the proposal
-
-        proposal = Proposal.objects.filter(team_lead=request.user.student, course_id=course.id).first()
-        if proposal is None:
-            proposal = Proposal()
-
-        print(course)
+        proposal = Proposal()
 
         proposal.title = request.POST.get("title")
         proposal.course = course
@@ -239,10 +223,56 @@ class ProposalSubmissionView(LoginRequiredMixin, View):
 
         proposal.save()
 
-        self.context = {
-            'students': Student.objects.all(),
+        return redirect('home')
+
+
+class ProposalUpdateView(LoginRequiredMixin, View):
+    template_name = 'project_management/proposal_update.html'
+
+    def get(self, request, *args, **kwargs):
+        pk = kwargs.get('id')
+        course = Course.objects.get(id=pk)
+        proposal_id = kwargs.get('proposal_id')
+        proposal = Proposal.objects.get(id=proposal_id)
+
+        context = {
+            'students': Student.objects.exclude(proposal__course=course),
             'teachers': Teacher.objects.all(),
             'proposal': proposal
         }
 
-        return redirect('home')
+        return render(request, self.template_name, context=context)
+
+    def post(self, request, *args, **kwargs):
+
+        proposal = Proposal.objects.get(id=kwargs.get('proposal_id'))
+
+        proposal.title = request.POST.get("title")
+        print(proposal.file)
+        proposal.save()
+        student_list = []
+
+        for student in json.loads(request.POST.get("tagify_student")):
+            student_list.append(student['value'])
+
+        students = Student.objects.filter(student_id__in=student_list)
+        proposal.students.set(students)
+
+        teacher_list = []
+
+        for teacher in json.loads(request.POST.get("tagify_teacher")):
+            teacher_list.append(teacher['value'])
+        teachers = Teacher.objects.filter(full_name__in=teacher_list)
+        proposal.preferred_supervisors.set(teachers)
+
+        messages.success(request, "Proposal has been updated")
+
+        proposal.save()
+
+        context = {
+            'students': Student.objects.exclude(proposal__course=proposal.course),
+            'teachers': Teacher.objects.all(),
+            'proposal': proposal
+        }
+
+        return render(request, self.template_name, context=context)
